test(geometry): pin mesh_weld's OOB guard at its exact boundary
`weld_indexed` documents that malformed input is treated as
"invalid-but-present rather than panicking". Its guard
`indices.iter().any(|&i| i as usize >= nverts)` mutated to `>` survives
9/9 — the off-by-one admits `i == nverts`, the smallest invalid index,
straight past the guard to panic at `remap[i]`.

The existing test could not catch it for two independent reasons, which
is why it looked covered: `out_of_range_index_is_a_no_op_not_a_panic`
uses index 9 against nverts 3 (far past the boundary, where `>` and `>=`
agree), AND its three vertices are all distinct, so the function returns
None via the "nothing merged" branch before ever reaching the
`remap[i as usize]` gather the guard protects.

The new fixture fixes both: two of three vertices share a position and
normal so a real merge occurs and the gather is reached, and one index
is exactly nverts. Under the mutant that panics with "index out of
bounds: the len is 3 but the index is 3".

LATENT — production feeds this indices produced by the mesher, so a
genuinely OOB index would signal an upstream bug. But the guard's own
contract is no-panic on malformed input, and it failed that at exactly
the boundary it exists for.

mesh_weld.rs hit 422 lines against the 400 budget, so its tests moved to
a sibling mesh_weld_tests.rs via `#[path]`, matching
contour_bool2d.rs/contour_bool2d_tests.rs. Now 194 lines; ratchet 4/4.

geometry 578 lib tests pass (was 577). clippy -D warnings clean on the
touched files.

Two negatives worth recording. `issue_2019_void_order_invariance.rs`
(actually in rust/processing) is NOT vacuous: it carries its own control
asserting the sequential CSG kernel IS order-sensitive, plus a check
that the cut changed the mesh at all, before asserting invariance across
a real byte reordering. And `coaxial_union.rs:377`'s slab test survives
strict-inequality mutation, but that is an equivalent mutant, not a gap:
`mid` is the strict midpoint of adjacent coalesced breakpoints and
`f.z_lo`/`f.z_hi` are themselves breakpoints, so equality is
structurally unreachable.
